requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
docker==7.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
requests==2.32.3
smmap==5.0.1
urllib3==2.2.2
